add MemPerGPUBufferGB parameter
Mostly we need this because our heuristics for choosing the GPU buffer size aren't very good. Too big, and pinning dominates the runtime. Too small, and the sim won't run.

+++ singlestep/Direct/DeviceFunctions.h
@@ -1,6 +1,6 @@
 #ifdef CUDADIRECT
 
-extern "C" void GetDeviceInfo(double *memGB, std::string &name);
+extern "C" void GetDeviceInfo(size_t *memGB, std::string &name);
 
 extern "C" int GetNGPU();
 extern "C" void GPUSetup(int cpd, uint64 MaxBufferSize,
